move /api/auth.js -> api.js
